requirements.txt
httpx>=0.27.0
google-genai>=1.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pyotp>=2.9.0
schedule>=1.2.0
